basic benchmark

+++ src/test/java/org/reactivity/benchmark/CountWordInFileBK.java
@@ -31,7 +31,7 @@ import org.reactivity.data.FileTestGenerator;
 * */
 public class CountWordInFileBK {
     private static final ActorSystem actorSystem = ActorSystem.create();
-    private static int nWord = FileTestGenerator.DefaultWordCount;
+    private static long nWord = FileTestGenerator.DefaultWordCount;
     private static String wordToCount = UUID.randomUUID().toString();
     private static File fileToTest;
